requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.14.0
requests==2.31.0
python-dotenv==1.0.0
openai==0.27.8
aiohttp==3.8.5
cachetools==5.3.1
pytz==2023.3
